+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/configuration/Locations.java
@@ -155,6 +155,7 @@ public class Locations {
 
             case SQL_INTERFACE:
                 return Constants.CLIENT_PACKAGE;
+
         }
         throw new UnsupportedOperationException("Unknown file " + file);
     }